// src/app/openapi/openapi-gen/model/contact.ts
/**
 * API for Contact App
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @springdoc.version@
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Email } from './email';
import { Address } from './address';
import { Phone } from './phone';
import { ContactGroup } from './contactGroup';
import { ContactRelation } from './contactRelation';


export interface Contact {
    id?: number;
    /**
     * The photo of the contact
     */
    photo?: string;
    /**
     * The thumbnail of the contact
     */
    thumbNail?: string;
    photoContentType?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    company?: string;
    birthDate?: string;
    notes?: string;
    phones?: Set<Phone>;
    addresses?: Set<Address>;
    emails?: Set<Email>;
    relations?: Set<ContactRelation>;
    groups?: Set<ContactGroup>;
}